Fix findGitRoot
The original code used git rev-parse --git-common-dir then did a .. from
it. This is incorrect when inside a git submodule (amongst other cases).
Change with --show-toplevel which is already an absolute path.

Partial fix for #182; this is insufficient, as now the container fails
to find the git repository `.git` directory